Create mixin from check_warnings in TestFloatingPointExceptions
The check_warnings context manager is now available for other
tests by inheriting from CheckWarningsMixin.
